Move test to related file.
